requirements.txt
Django
django_rest_passwordreset
djangorestframework
djangorestframework_simplejwt==5.3.0
django-cors-headers
drf-yasg
requests

